LaCarteAuxTresors/src/main/java/org/carbon/model/Orientation.java
package org.carbon.model;

public enum Orientation {
    NORTH, EAST, SOUTH, WEST;

    public static Orientation mapOrientation(String orientation) {
        switch (orientation) {
            case "N":
                return NORTH;
            case "E":
                return EAST;
            case "S":
                return SOUTH;
            case "O":
                return WEST;
            default:
                return null;
        }
    }

    public Orientation turnRight() {
        switch (this) {
            case NORTH:
                return EAST;
            case EAST:
                return SOUTH;
            case SOUTH:
                return WEST;
            case WEST:
                return NORTH;
            default:
                return null;
        }
    }

    public Orientation turnLeft() {
        switch (this) {
            case NORTH:
                return WEST;
            case EAST:
                return NORTH;
            case SOUTH:
                return EAST;
            case WEST:
                return SOUTH;
            default:
                return null;
        }
    }

    @Override
    public String toString() {
        switch (this) {
            case NORTH:
                return "N";
            case EAST:
                return "E";
            case SOUTH:
                return "S";
            case WEST:
                return "O";
            default:
                return null;
        }
    }
}
